refactor to replace AgentPubKey with AgenPubKeyB64
As per Guillem's advice, using AgentPubKey as a HashMap key (see
the PlayerStats type) in JS breaks things as JS isn't ready to
have binary array as HashMap key.
Instead we need to use a binary string AgentPubKeyB64.
This commit replaces the usage with some very adhoc code and also
introduces draft of the convert_keys_from_b64 to use for that.